add support for exotic result notation - closes #14249
If a check-in time is set for a tournament and the opponent does not show up,
the opponent who is not there loses.
However, since no game was played here, this is indicated with + and -.
If it says +:-, for example, it means that white has won but has not played.
With -:+ it is the other way around. Black gets the point but has not played.
If neither player comes, the game is rated -:-.
However, I have only experienced this once before at a major chess tournament.
https://github.com/lichess-org/lila/issues/14249